+++ PKHeX.Core/Legality/Encounters/Generator/Possible/EncounterPossible7.cs
@@ -231,7 +231,7 @@ public record struct EncounterPossible7(EvoCriteria[] Chain, EncounterTypeGroup 
         for (; Index < db.Length;)
         {
             var enc = db[Index++];
-            if (!enc.CanBeReceivedByVersion(Version))
+            if (!enc.CanBeReceivedByVersion((int)Version))
                 continue;
             foreach (var evo in Chain)
             {
